fix(courtcase): keep the case-number/court heading, not the party cause title
PR #251 made the court-case detail heading a human-readable cause title
"<plaintiff> विरुद्ध <defendant>" and demoted the case-number/court identifier
to a muted sub-line. On real dockets the party strings are enormous: e.g.
supreme/075-wf-0005 renders a ~40-word heading, so the page loads showing a
clean "075-WF-0005 (सर्वोच्च अदालत)" and then flips to that wall of text once
the data arrives.

Keep the identifier as the heading (what is already shown before load). The
parties still appear in the dedicated Plaintiff/Defendant row below, so no
information is lost. This reverts the cause-title heading, the tab title /
JSON-LD name, the getCourtCaseCauseTitle helper, and the courtVersus i18n key.

The actual regression fix from #251 is preserved: CourtCaseProfile still passes
the canonical @id IRI to CourtCaseCard, so the identifier renders as
"075-WF-0005 (Supreme Court)" instead of the raw "supreme:075-wf-0005".
