+++ src/i18n/locales/en.json
@@ -599,7 +599,6 @@
     "courtHearing": "Hearing",
     "courtPlaintiff": "Plaintiff",
     "courtDefendant": "Defendant",
-    "courtVersus": "v.",
     "courtHearings": "Hearings",
     "courtHearingDate": "Hearing Date",
     "courtHearingStatus": "Status",
