+++ src/i18n/locales/ne.json
@@ -597,7 +597,6 @@
     "courtHearing": "सुनुवाइ",
     "courtPlaintiff": "वादी",
     "courtDefendant": "प्रतिवादी",
-    "courtVersus": "विरुद्ध",
     "courtHearings": "सुनुवाइहरू",
     "courtHearingDate": "सुनवाइ मिती",
     "courtHearingStatus": "मुद्दाको स्थिती",
