+++ src/pages/CourtCaseProfile.tsx
@@ -1,6 +1,5 @@
 import { useParams, Link } from "react-router-dom";
 import { Helmet } from "react-helmet-async";
-import { useTranslation } from "react-i18next";
 import { useQuery } from "@tanstack/react-query";
 import { AlertCircle, ArrowLeft } from "lucide-react";
 
@@ -8,7 +7,6 @@ import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { CourtCaseCard } from "@/components/CourtCaseCard";
 import { getCourtCaseFull } from "@/services/datalake-api";
-import { getCourtCaseCauseTitle } from "@/utils/courtCase";
 
 // The /courtcase/* splat tail is the courtcase IRI path component
 // `<court>/<case_number>` (e.g. `special/081-CR-0079`); we rebuild the canonical
@@ -26,7 +24,6 @@ function parseTail(tail: string): { court: string; caseNumber: string } | null {
 }
 
 export default function CourtCaseProfile() {
-  const { t } = useTranslation();
   const params = useParams();
   const tail = params["*"] || "";
   const parsed = parseTail(tail);
@@ -46,15 +43,8 @@ export default function CourtCaseProfile() {
   const courtCaseIri = parsed
     ? `https://jawafdehi.org/courtcase/${parsed.court.toLowerCase()}/${parsed.caseNumber.toLowerCase()}`
     : "";
-  // Prefer the human-readable cause title ("<plaintiff> विरुद्ध <defendant>") for
-  // the tab title / JSON-LD name; fall back to the bare case number when the
-  // parties are unknown.
-  const causeTitle = data
-    ? getCourtCaseCauseTitle(data, t("caseDetail.courtVersus", "v."))
-    : null;
-  const heading = causeTitle || caseNumber;
   const title = data
-    ? `${heading} — ${data.case_type || "Court case"}`
+    ? `${caseNumber} — ${data.case_type || "Court case"}`
     : caseNumber || "Court case";
 
   // schema.org JSON-LD for crawlers (parity with the retired R2 landing pages). A
